+++ src/cltk/prosody/lat/hexameter_scanner.py
@@ -16,13 +16,13 @@ import re
 
 from Levenshtein import distance
 
-import cltk.prosody.latin.string_utils as string_utils
-from cltk.prosody.latin.metrical_validator import MetricalValidator
-from cltk.prosody.latin.scansion_constants import ScansionConstants
-from cltk.prosody.latin.scansion_formatter import ScansionFormatter
-from cltk.prosody.latin.syllabifier import Syllabifier
-from cltk.prosody.latin.verse import Verse
-from cltk.prosody.latin.verse_scanner import VerseScanner
+import cltk.prosody.lat.string_utils as string_utils
+from cltk.prosody.lat.metrical_validator import MetricalValidator
+from cltk.prosody.lat.scansion_constants import ScansionConstants
+from cltk.prosody.lat.scansion_formatter import ScansionFormatter
+from cltk.prosody.lat.syllabifier import Syllabifier
+from cltk.prosody.lat.verse import Verse
+from cltk.prosody.lat.verse_scanner import VerseScanner
 
 __author__ = ["Todd Cook <todd.g.cook@gmail.com>"]
 __license__ = "MIT License"
