+++ src/cltk/prosody/lat/metrical_validator.py
@@ -7,7 +7,7 @@ from typing import List
 
 from Levenshtein import distance
 
-from cltk.prosody.latin.scansion_constants import ScansionConstants
+from cltk.prosody.lat.scansion_constants import ScansionConstants
 
 LOG = logging.getLogger(__name__)
 LOG.addHandler(logging.NullHandler())
