+++ tests/test_lemmatize.py
@@ -47,7 +47,7 @@ class TestSequenceFunctions(unittest.TestCase):
             ("componam", "UNK"),
         ]  # pylint: disable=line-too-long
         jv_replacer = JVReplacer()
-        tokenizer = WordTokenizer("latin")
+        tokenizer = WordTokenizer("lat")
         test_str = test_str.lower()
         test_str = jv_replacer.replace(test_str)
         tokens = tokenizer.tokenize(test_str)
@@ -65,7 +65,7 @@ class TestSequenceFunctions(unittest.TestCase):
             ("componam", "componam"),
         ]  # pylint: disable=line-too-long
         jv_replacer = JVReplacer()
-        tokenizer = WordTokenizer("latin")
+        tokenizer = WordTokenizer("lat")
         test_str = test_str.lower()
         test_str = jv_replacer.replace(test_str)
         tokens = tokenizer.tokenize(test_str)
@@ -89,7 +89,7 @@ class TestSequenceFunctions(unittest.TestCase):
             ("componam", "compono"),
         ]  # pylint: disable=line-too-long
         jv_replacer = JVReplacer()
-        tokenizer = WordTokenizer("latin")
+        tokenizer = WordTokenizer("lat")
         test_str = test_str.lower()
         test_str = jv_replacer.replace(test_str)
         tokens = tokenizer.tokenize(test_str)
@@ -115,7 +115,7 @@ class TestSequenceFunctions(unittest.TestCase):
             ("componam", "compono"),
         ]  # pylint: disable=line-too-long
         jv_replacer = JVReplacer()
-        tokenizer = WordTokenizer("latin")
+        tokenizer = WordTokenizer("lat")
         test_str = test_str.lower()
         test_str = jv_replacer.replace(test_str)
         tokens = tokenizer.tokenize(test_str)
@@ -129,7 +129,7 @@ class TestSequenceFunctions(unittest.TestCase):
         test_str = "amabimus"
         target = [("amabimus", "amo")]
         jv_replacer = JVReplacer()
-        tokenizer = WordTokenizer("latin")
+        tokenizer = WordTokenizer("lat")
         test_str = test_str.lower()
         test_str = jv_replacer.replace(test_str)
         tokens = tokenizer.tokenize(test_str)
@@ -194,7 +194,7 @@ class TestSequenceFunctions(unittest.TestCase):
             ("componam", "compono"),
         ]  # pylint: disable=line-too-long
         jv_replacer = JVReplacer()
-        tokenizer = WordTokenizer("latin")
+        tokenizer = WordTokenizer("lat")
         test_str = test_str.lower()
         test_str = jv_replacer.replace(test_str)
         tokens = tokenizer.tokenize(test_str)
@@ -228,7 +228,7 @@ class TestSequenceFunctions(unittest.TestCase):
             ("componam", "compono", "<DictLemmatizer: Morpheus Lemmas>"),
         ]  # pylint: disable=line-too-long
         jv_replacer = JVReplacer()
-        tokenizer = WordTokenizer("latin")
+        tokenizer = WordTokenizer("lat")
         test_str = test_str.lower()
         test_str = jv_replacer.replace(test_str)
         tokens = tokenizer.tokenize(test_str)
